10/19 spots working

+++ backend/db/migrations/20250314210843-create-spot.js
@@ -45,11 +45,11 @@ module.exports = {
         allowNull: false,
       },
       lat: {
-        type: Sequelize.DECIMAL(10,7),
+        type: Sequelize.FLOAT(10,7),
         allowNull: false,
       },
       lng: {
-        type: Sequelize.DECIMAL(10,7),
+        type: Sequelize.FLOAT(10,7),
         allowNull: false,
       },
       name: {
